pyproject.toml
[project]
name = "ai-content-processor"
version = "1.0.0"
description = "AI-powered audio transcription, image analysis, and document processing"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    {name = "Muhammad Omar Muneer", email = "mumarm45@gmail.com"}
]
license = {text = "MIT"}
keywords = ["ai", "transcription", "image-analysis", "claude", "whisper", "nlp"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "anthropic>=0.40.0",
    "faster-whisper==1.0.3",
    "gradio>=5.0.0",
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "numpy<2",
    "pillow>=10.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "torch>=2.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
ai-processor = "src.web.app:launch_app"

[build-system]
requires = ["setuptools>=65.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
ai_processor = ["py.typed"]

[tool.black]
line-length = 100
target-version = ["py312"]
include = '\.pyi?$'
extend-exclude = '''
/(
  \.git
  | \.venv
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 100
target-version = "py312"
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=ai_processor --cov-report=html"
